refactor: clean up VM module
