+++ src/vector_store.rs
